Refactor project structure into src packages
