+++ app.py
@@ -5,8 +5,15 @@ This file imports and registers all handlers from the frontend module.
 Run with: chainlit run app.py
 """
 
-# Import all handlers and actions to register them with Chainlit
-from frontend import (
+import sys
+from pathlib import Path
+
+PROJECT_ROOT = Path(__file__).resolve().parent
+SRC_PATH = PROJECT_ROOT / "src"
+if str(SRC_PATH) not in sys.path:
+    sys.path.insert(0, str(SRC_PATH))
+
+from src.frontend import (
     on_chat_start,
     on_settings_update,
     on_message,
