+++ src/backend/config.py
@@ -1,0 +1,20 @@
+"""Shared configuration constants for backend modules."""
+import os
+
+DEFAULT_DATA_DIR = "CSV"
+NBA_DEFAULT_SEASON = os.environ.get("NBA_API_DEFAULT_SEASON", "2023-24")
+ENTITY_SAMPLE_SIZE = 1000
+SEARCH_SAMPLE_SIZE = 1000
+CSV_EXECUTION_TIMEOUT = 30
+API_EXECUTION_TIMEOUT = 60
+CHART_HISTORY_LIMIT = 10
+CHART_ROW_LIMIT = 10
+RAW_RESULT_TRUNCATION = 5000
+ANALYSIS_TRUNCATION = 2000
+EXEC_RESULT_TRUNCATION = 3000
+"""
+Constants governing prompt construction.
+"""
+MAX_PLAN_STEPS = 6
+MIN_PLAN_STEPS = 4
+SUCCESSFUL_PATTERN_LIMIT = 10
