+++ src/backend/nodes/__init__.py
@@ -1,0 +1,45 @@
+"""
+Public exports for backend node classes and shared utilities.
+"""
+
+from backend.nodes.analysis import DeepAnalyzer, ResponseSynthesizer, Visualizer
+from backend.nodes.code_generation import CodeGenerator, NBAApiCodeGenerator
+from backend.nodes.data_ingestion import DataMerger, LoadData, NBAApiDataLoader
+from backend.nodes.entity import EntityResolver, SearchExpander
+from backend.nodes.execution import ErrorFixer, Executor, SafetyCheck
+from backend.nodes.planning import ContextAggregator, Planner
+from backend.nodes.query import AskUser, ClarifyQuery
+from backend.nodes.schema import DataProfiler, SchemaInference
+from backend.nodes.validation import CrossValidator, ResultValidator
+from backend.utils.call_llm import call_llm
+from backend.utils.data_source_manager import data_source_manager
+from backend.utils.knowledge_store import knowledge_store
+from backend.utils.nba_api_client import nba_client
+
+__all__ = [
+    "LoadData",
+    "NBAApiDataLoader",
+    "DataMerger",
+    "SchemaInference",
+    "DataProfiler",
+    "ClarifyQuery",
+    "AskUser",
+    "EntityResolver",
+    "SearchExpander",
+    "ContextAggregator",
+    "Planner",
+    "CodeGenerator",
+    "NBAApiCodeGenerator",
+    "SafetyCheck",
+    "Executor",
+    "ErrorFixer",
+    "ResultValidator",
+    "CrossValidator",
+    "DeepAnalyzer",
+    "Visualizer",
+    "ResponseSynthesizer",
+    "call_llm",
+    "data_source_manager",
+    "knowledge_store",
+    "nba_client",
+]
